+++ packages/frontend/src/pages/designs/index.vue
@@ -30,7 +30,6 @@
 
 <script setup lang="ts">
 import { ProjectTypeScope, type ProjectType } from '#imports';
-import { sortBy, uniq } from 'lodash-es';
 
 definePageMeta({
   title: 'Designs',
@@ -48,14 +47,10 @@ const importBtnRef = useTemplateRef('importBtnRef');
 
 const listViewRef = useTemplateRef('listViewRef');
 const statusOptions = computed(() => apiSettings.settings?.statuses?.map(status => ({title: status.label, value: status.id, icon: status.icon})) || []);
-const suggestedTags = ref<string[]>([]);
-watch(() => listViewRef.value?.items?.data.value as ProjectType[]|undefined, (items) => {
-  if (!items) { return; }
-  suggestedTags.value = sortBy(uniq(items.flatMap(p => p.tags).concat(suggestedTags.value)));
-}, { immediate: true, deep: 1 });
+const suggestedTags = useProjectTypeTags();
 const filterProperties = computed((): FilterProperties[] => [
   { id: 'status', name: 'Status', icon: 'mdi-flag', type: 'select', options: statusOptions.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
-  { id: 'tag', name: 'Tag', icon: 'mdi-tag', type: 'combobox', options: suggestedTags.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
+  { id: 'tag', name: 'Tag', icon: 'mdi-tag', type: 'combobox', options: suggestedTags.getTags, allow_exclude: true, allow_regex: false, default: '', multiple: true },
   { id: 'timerange', name: 'Time Created', icon: 'mdi-calendar', type: 'daterange', options: [], allow_exclude: true, default: '', multiple: true },
   { id: 'language', name: 'Language', icon: 'mdi-translate', type: 'select', options: apiSettings.settings!.languages.map(l => l.code), allow_exclude: true, default: '', multiple: true },
 ]);