+++ packages/frontend/src/pages/projects/finished.vue
@@ -42,15 +42,14 @@ const apiSettings = useApiSettings();
 
 const listViewRef = useTemplateRef('listViewRef');
 const suggestedMembers = ref<string[]>([]);
-const suggestedTags = ref<string[]>([]);
 watch(() => listViewRef.value?.items?.data.value as PentestProject[]|undefined, (items) => {
   if (!items) { return; }
   suggestedMembers.value = sortBy(uniq(items.flatMap(p => p.members.map(member => member.username)).concat(suggestedMembers.value)));
-  suggestedTags.value = sortBy(uniq(items.flatMap(p => p.tags).concat(suggestedTags.value)));
 }, { immediate: true, deep: 1 });
+const suggestedTags = useProjectTags();
 const filterProperties = computed((): FilterProperties[] => [
   { id: 'member', name: 'Member', icon: 'mdi-account', type: 'combobox', options: suggestedMembers.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
-  { id: 'tag', name: 'Tag', icon: 'mdi-tag', type: 'combobox', options: suggestedTags.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
+  { id: 'tag', name: 'Tag', icon: 'mdi-tag', type: 'combobox', options: suggestedTags.getTags, allow_exclude: true, allow_regex: false, default: '', multiple: true },
   { id: 'timerange', name: 'Time Created', icon: 'mdi-calendar', type: 'daterange', options: [], allow_exclude: true, default: '', multiple: true },
   { id: 'language', name: 'Language', icon: 'mdi-translate', type: 'select', options: apiSettings.settings!.languages.map(l => l.code), allow_exclude: true, default: '', multiple: true },
 ]);