+++ packages/frontend/src/pages/templates/index.vue
@@ -44,7 +44,7 @@
 </template>
 
 <script setup lang="ts">
-import { sortBy, uniq, capitalize } from 'lodash-es';
+import { capitalize } from 'lodash-es';
 
 definePageMeta({
   title: 'Templates',
@@ -102,15 +102,11 @@ async function performCreate() {
 const statusOptions = computed(() => apiSettings.settings?.statuses?.map(status => ({title: status.label, value: status.id, icon: status.icon})) || []);
 const languageOptions = computed(() => apiSettings.settings!.languages.map(l => ({title: l.name, value: l.code, icon: 'mdi-translate'})));
 const riskLevelOptions = computed(() => Object.values(RiskLevel).map(l => ({title: capitalize(l), value: l})));
-const suggestedTags = ref<string[]>([]);
-watch(() => listViewRef.value?.items?.data.value as FindingTemplate[]|undefined, (items) => {
-  if (!items) { return; }
-  suggestedTags.value = sortBy(uniq(items.flatMap(p => p.tags).concat(suggestedTags.value)));
-}, { immediate: true, deep: 1 });
+const suggestedTags = useFindingTemplateTags();
 const filterProperties = computed((): FilterProperties[] => [
   { id: 'status', name: 'Status', icon: 'mdi-flag', type: 'select', options: statusOptions.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
   { id: 'risk_level', name: 'Risk Level', icon: 'mdi-alert', type: 'select', options: riskLevelOptions.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
-  { id: 'tag', name: 'Tag', icon: 'mdi-tag', type: 'combobox', options: suggestedTags.value, allow_exclude: true, allow_regex: false, default: '', multiple: true },
+  { id: 'tag', name: 'Tag', icon: 'mdi-tag', type: 'combobox', options: suggestedTags.getTags, allow_exclude: true, allow_regex: false, default: '', multiple: true },
   { id: 'timerange', name: 'Time Created', icon: 'mdi-calendar', type: 'daterange', options: [], allow_exclude: true, default: '', multiple: true },
   { id: 'language', name: 'Language', icon: 'mdi-translate', type: 'select', options: languageOptions.value, allow_exclude: true, default: '', multiple: true },
 ]);